pyproject.toml
[tool.poetry]
name = "netpred"
version = "0.1.0"
description = "An interpretable model for secondary structure prediction"
authors = ["James Robinson <robioson@pm.me>"]

[tool.poetry.dependencies]
python = ">=3.9"
biopython = "1.77"
torch = "^1.9.1"
captum = "^0.5.0"
notebook = "^6.4.8"
matplotlib = "^3.5.1"
numpy = "^1.22.1"
seaborn = "^0.11.2"
scikit-learn = "^1.0.2"
peptides = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
